src/main/java/com/sparta/engineering50/WindPOJO.java
package com.sparta.engineering50;

public class WindPOJO {
    double speed;
    double deg;
    double gust;

    public WindPOJO(){}

    public WindPOJO(double speed, double deg, double gust) {
        this.speed = speed;
        this.deg = deg;
        this.gust = gust;
    }

    public double getSpeed() {
        return speed;
    }

    public void setSpeed(double speed) {
        this.speed = speed;
    }

    public double getDeg() {
        return deg;
    }

    public void setDeg(double deg) {
        this.deg = deg;
    }

    public double getGust() {
        return gust;
    }

    public void setGust(double gust) {
        this.gust = gust;
    }
}
